SecureVision_LPR/lpr_backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-env==0.2.1
django-environ==0.4.5
dnspython==1.16.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
imageio==2.9.0
imutils==0.5.3
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pipenv==2018.11.26
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2020.7.4
torch==1.5.1
torchvision==0.6.1
# torch-scatter==1.4.0
# torch-sparse==0.4.3
# torch-cluster==1.4.5
virtualenv==20.0.27
virtualenv-clone==0.5.4
